requirements.txt
fastapi
uvicorn
python-telegram-bot==21.6
init-data-py